virt: Ignore 3 unattended_install serial log errors
Previously it was possible due to timing, for the test
to clobber the vm's serial log file, then monitor
the wrong file.  This patch makes it ignore the first
three attempts to read the file, then it starts making noise.
There's no need I'm aware of why the test should have
to ever create the log file, it just needs to be more
patient.

Signed-off-by: Chris Evich <cevich@redhat.com>