libvirt-virsh: Get virsh_exec on remote for VirshConnectBack class.
Sometimes, the virsh command path on local host is different with
it on remote host. E.g, if libvirt on local is hand-builded, os_dep('virsh')
will return /usr/local/bin/virsh. But if libvirt on remote is installed with
libvirt.rpm, /usr/local/bin/virsh is not accessable with no such file. It
exists at /usr/bin/virsh.

Signed-off-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>