virttest.utis_env: Prevent to raise 'RuntimeError' during dumping the env
The env dictionary has some nested dictionaries accessible by other
threads or precesses and sometimes it causes an 'RuntimeError' that
dictionary size has been changed. One known way is that the 'tcpdump'
process adds a new mac->ip record into the env["address_cache"]
dictionary and the 'pickle.dump' than fails with the 'RuntimeError'
exception.

By this patch we will create a copy of a dictionary and after that
copy of nested dictionaries to prevent this error.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>