virttest.env_process: Remove tcpdump from environment
Pavel found a bug in which the tcpdump thread, stored
in the environment class, writes a new DHCP lease at
the moment the environment is being saved (pickled),
causing a pickle error, given that the dict size of
the "address_cache" dict changed during the pickling
process.

After a lot of discussion, I came to the conclusion
the best way to avoid this problem is to stop the
thread before saving the environment, any DHCP lease
happening post post-processing should not be recorded
anyway. The reason why tcpdump was being stored in
the environment was basically because it is possible
to re-use the process among tests, due to aexpect.

Since we don't *really* need tcpdump to persist
between tests, we don't need it stored in the
environment as well. We can just turn it into another
thread, such as the screendump thread, and make sure
it is always ended in the postprocessing stage. This
should solve the bug and make the env structure simpler,
not requiring extra hacks. In fact, this allows us to
cut one hack at the function that destroys the
environment.

CC: Pavel Hrdina <phrdina@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>