virt: change test entry point method name
There's no need to have the name of the "run_<name>" method (main
test entry point) depend on the test name. So, for simplicity sake,
let's just name the test entry point "run".

I have decided to not fallback gracefully to the previous naming
convention as I believe this can lead to confusion in the future.

Signed-off-by: Cleber Rosa <crosa@redhat.com>