virttest.standalone_test: Shut down all VMs at the end of ./run
In the early days, it was thought that keeping QEMU processes
up and running after each test suite execution was good to make
testing faster. In practice, it only confuses people to see that
after the suite is done, there are still a lot of aexpect
processes hanging around.

So let's just destroy the environment at the end of the test,
trying to shut down all vms in there nicely, so we won't have
any leftovers after ./run is done.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>