virttest.qemu_vm: Bugfix - make seabios file to be deleted on destroy
Turns out we were clearing self.logs even when there was
a previous vm state. This made the seabios file to not
be deleted during vm cleanup. Let's fix this by moving
initialization of the variable to the right block.

With this, now virt-test won't leave tons of /tmp/seabios*
sockets around even when it is done.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>