virt.virtnet : update vm.virtnet when guest reboot
When starting a guest with one nic device, then
starting a guest with two nics (not removing env),
the second nic can not be initialized correctly.

When booting a guest with nic_extra_params, then
booting another guest without this option, the
second guest will boot with nic_extra_params
unexpectedly.

The root cause of these issues is that we don't
update virtnet when the VM object is preprocessed.

Signed-off-by: Yunping Zheng <yunzheng@redhat.com>