virt: Reduce virtnet and rewrite unittests
Component classes of virtnet had their behavior bound together,
making them less flexible and more difficult to use.  An auxiliary
class was then used to support customization but it was not at the
right level in the stack, nor did it help untangle any bound
behavior.

This update takes the intention of the auxiliary class
(VMNetStyle) and folds it into the virtnet component classes
in a more heirarchical manner.  Additionally, VirtNet itself
has been split up into an ABC, and multiple subclasses which
operate on different data sources, but through a common interface.

This places the burden/responsibility for managing data and it's source
on the virtnet consumer.  It allows for greater flexibility and eases
maintenance since class responsibility and interfaces are more clearly
defined and don't come with any cross-dependencies.

Finally, since a prior update made PropCan instances fully pickleable,
there is no need to perform string conversion/parsing for storage.
Further, since a proper ABC model was implemented, storage can
be performed agnostic to inner container-class type.  These changes
permit rich comparison between the various VirtNet classes, including
the ability to select specific properties to NOT be compared, at
runtime.

The unittests were nearly completely re-written to cover the most
critical points in the new structure.  Test coverage is a little
less robust than previous, however it also runs roughly 10x faster.
The cartesian-based testing was updated to reflect recent changes
to the vm classes, and cut back on a lot of redundant tests.

Signed-off-by: Chris Evich <cevich@redhat.com>