qemu.tests.migration: Shut down VM if killing migration bg cmd times out
Sometimes, post migration run, we can run into a condition
where tcpdump or the VM hangs at migration cleanup stage.
After analyzing test logs, including VM logs, there's nothing
that points out to any guest OS crash or similar condition.

With that in mind, it is admissible to be more lenient
when trying to kill the migration background command, and
if a timeout happens when trying to kill it, be safe and
try to shut down the vm gracefully.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>