virt: Simplify needs_restart()
It's impossible to perfectly track vm state from test to test.
This area / behavior is currently being discussed for further
refinement, for now rather than try and be smart, just assume
a restart is needed.

The thinking is: The consequences of a wrong restart determination
are more costly in the long-run than the time-savings gained by
not restarting the qemu process.  Really this decision needs more
refinement inside env_process anyway, not the BaseVM class.

Signed-off-by: Chris Evich <cevich@redhat.com>