virsh_create_pool: Fix test error
The test tried to use a non-existent "/tmp/<file>.xml" - so instead
of doing that - I've created a "tmp/<file>.xml" on the fly using the
data images directory as the pool location (e.g. where the jeos lives)

Secondarily, I found that the test was destroying my running default
pool. So two subsequent runs would have different results. Turns out
there is a positive and negative run for "default" pool copying - in
one instance the pool isn't stopped and a new one is attempted to be
created, resulting in the expected error; however, in the finally path
there was no check to not destroy the "created" pool.

Finally, the deletion & recreation of the default pool missed resetting
the pool 'autostart' attribute. Taking a cue from the net-* code which
had similar functionality - add code that will save off the autostart state
of each pool so that it can be regenerated afterwards.