run: Fix logic in new run options
When an option is set, this option will override that is defined in the
cfg file. Otherwise, the behavior is defined by the cfg file.

So when processing options, we should not use negative condition 'if not',
because this will impose the option when it's not set and follow the cfg
file when set. This is not the hehavior we expect.

To set the default behavior, we need to change the shared base cfg file.

Signed-off-by: Hao Liu <hliu@redhat.com>