virt: Fix host IP lookup assumption
Looking up an IP address for the host system should not depend
on guest networking configuration.  For example, if guest is
using nettype=user, then (technically) the netdst param contents
are not applicable and may not even be set!  However,
utils_net.get_host_ip_address() assumes netdst param will
always point at a host networking device.

Another problematic situation is when a guest is installed
with unattended_install using a network-based method,
however guest is initially configured WITHOUT any networking
device (possible with libvirt/virt-install).

A safer assumption, is that a default route is defined to
deliver packets via a valid local networking device.  With IPv4,
this networking device _must_ have an associated IP address
which is unlikely to be invalid or a black-hole.  Therefor,
this address is a much safer option than relying on guest
networking configuration.

Signed-off-by: Chris Evich <cevich@redhat.com>