run: Add --update-config
Replace the script update_config.py with the option
--update-config to the runner script. It will simply
update all config files, overriding any manual configs
and/or files that had upstream changes.

The reason we don't do it compulsorily is because we
find it is rude to just override other people's
customizations, and there's no obvious way to tell
when it's simply an upstream update or someone that
edited one particular file.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>