virt: Support _min _max _equal suffix in configure.
After this patch, A_min set minimum for A, A_max set maximum for A,
A_equal set value for A.
This feature is useful when we set a parameter and do not want it covered by
new value late.

e.g:
mem_min = 1024
mem = 2048

Then mem = 1024 will be used.

Signed-off-by: Feng Yang <fyang@redhat.com>