qemu.tests.virtio_console: Force new params in vm.create()
This test started with "start_vm = no" and we are starting the VM
manually. It's not guaranteed, that params were updated in preprocess
so we need to make sure params were updated by calling create with
new params.

One example of incorrect params is when VM exists and is alive from
previous test. In such case preprocess skips initialization (start_vm
= no") and vm.create() destroys and recreates the VM from previous test.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>