virttest.qemu_virtio_port: Check booth size and presence of guest script
This patch adds additional check for compiled guest script file. This is
necessarily because it might happend, that the script was copied/already
exists in the guest machine and is of the correct version, but wasn't
successfully compiled. Originally this wouldn't be dettect and compilation
won't be executed, thus all following tests fail.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>