test.ethtool: Disable large frame check when gso on
In old kernel will do tcpdump before gso split the big packet,
so we can capture the big packets via tcpdump.
Currently when enable gso, tcpdump only capture the packets
after gso split the big packet, so the tcpdump can not capture
big packet. this is not a suitable method for gso, disable it.

Signed-off-by: Yunping Zheng <yunzheng@redhat.com>