qemu.test: Update the check method in smbios_table
In the machine_type check for smbios_table, we only care the main version
information but not the details. And the command line output from guest
will include more than the version info:
System.Version mismatch, Set 'RHEL 7.0.0 PC'
but guest is : 'RHEL 7.0.0 PC (i440FX + PIIX, 1996)'

So update the check method for it.

Signed-off-by: Yiqiao Pu <ypu@redhat.com>