virttest: Move sessions from virt_vm to specific vm class
Move sessions from virt_vm to specific vm class. And dont clean it
when clone status is True. So when we clone a vm, these objects can
be saved in new vm object. Otherwise when we try to destroy a cloned
vm object. Many sessions will be left in system.

Signed-off-by: Yiqiao Pu <ypu@redhat.com>