virttest.standalone_test: If -c used, tags are shortnames
When using -c, all bets about how the suite is used are
off, as for the number of guests used and other details.

So, to avoid name collisions, ensure the test tag in case
we're using our own configuration is equal to the shortname,
not the special, trimmed down tag we use when not using -c.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>