virttest.libvirt_xml: Move load_python_module to base
The deceptively generic function name load_python_module
hid the fact that this was an API with an application
way too narrow to be in virttest.utils_test. So move it
back to the virttest.libvirt_xml module where it belongs.

Also, change the API name slightly to load_xml_module.
This is not a generic python module loader function,
it has a very specific application.

CC: Wayne Sun <gsun@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>