utils_misc: Update mount check function and fix a wrong arg in umount
When checking mountpoints, the nfs type displays 'nfs4' if mount
version 4 NFS server, so use a regex to search the mount items.
NFS not mounted is not an error, so change the message to debug
level.

Remove umount step as it doesn't make sense to umount in a mount
function.

In the umount function, the second arg in utils.system() is timeout,
so fix it as well. Also, disable verbose by default for
mount/umount/is_mount functions.

Signed-off-by: Yanbing Du <ydu@redhat.com>