Allow fetching results from multiple guest Autotest runs during one virt test.
This is one of the upgrades necessary for running Autotest on
multiple guests and expanding the original idea of running only
on one main guest (main_vm). There are use cases (e.g. a networking
test) where more than one virtual machines are necessary. In such
cases the guest_autotest_results folder is overwritten by each
following Autotest run on a guest and as a result the folder
contains only the results from the last guest. This patch allows
for generating multiple guest result folders ordered by time of
execution.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>