Update network configuration together with all other parameters on QEMU VM creation.
The expected VM behavior is that with the creation of a VM, its
parameters will be updated. This is indeed how it works, however,
with the exception of network configuration which for some reasons
remains the old one and can be updated only if the entire VM object
is reinitialized.

Since in a virt test the VM is generally retrieved from the
environment, we should keep this well defined practice and avoid
reinitializing the VM object ourselves.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>