Move any libvirt related warnings that appear even when using solely qemu to a more appropriate place.
The warnings about missing libvirtd come from a check during
importing of the utils_libvirtd module. The module is imported
in env_process and therefore the warnings appear regardless of
whether we use libvirt or qemu. With this commit, the check will
still be performed but the warning will appear only if the user
actually intends to use the libvirtd control system.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>