Add a parameter to make using a custom path for the env file optional.
The parameter 'other_tests_dirs' is already used for defining
alternative directory for virt tests and using it as a custom
path for the env file is ambiguous. It also forces the file in
that directory for every user with custom virt tests. This commit
separates the parameters allowing for custom env directory and
custom virt tests directory.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>