shared/cfg/guest-hw.cfg: Consolidate image_backend variant
It was brought to my attention that, when one passes
--image-type=raw to the test runner, we end up
not filtering out lvm related variants. Of course, that
is not what we want (when we use image_type = raw, we
are mainly assuming a raw discrete image inside a filesystem,
image.raw, if you will). This will cause boot lvm variants
to be executed, and as people do not have proper setup
to do it they will be confused as to why that is happening.

I also noticed that below the images_backend variant,
we do have a lvm related variant, and its indentation
level was wrong. I assume it is reasonable we merge
both variants and fix the indentation level of
the lvm related blocks.

This should fix people seeing lvm related variants
using --image-type=raw, while keeping the default
suite behavior not altered.

CC: Xu Tian <xutian@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>