numa: support to pin threads to extra node
In the past, we only try to pin threads to the cpus in one node.
If there is no enough nodes, we will try to pin a pair of vhost
thread and vcpu thread in same cpu. But if the number of vhost
threads isn't same as the number of vcpu threads, the pinning
isn't balanced.

This patch added support to pin threads to extra node.

Signed-off-by: Amos Kong <akong@redhat.com>