virttest.bootstrap: Add get_jeos_info() API
With the newest JeOS 20, people running libvirt
tests on earlier versions of Fedora as a host
will have trouble with the fact that virt-install
on those hosts will not know anything about
'fedora.20' guests.

The workaround existing in current virt-test,
use -g JeOS.19.x86_64.i440fx to select JeOS 19
and tools/download_manager.py to download the
right JeOS image is way too cumbersome for people
using a Fedora version that is still alive and
supported.

So let's use the latest JeOS only for people
running on Fedora 20, and leave the older JeOS 19
for people running on other hosts.

In order to accomplish this goal, add get_jeos_info(),
a function that will return JeOS information according
to the host OS info, and change references to the
JeOS variants and asset names to use that API.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>