utils_test/libvirt/get_all_cells(): Fix call to virsh.freecell
Commit id '30073c37' added a 'cellno' parameter to virsh.freecell as
the first optional/named parameter; however, the call here didn't indicate
or name whether its first parameter was for 'options', thus the assumption
was it was for cellno.  This led to test failure in 'virsh.nodememstats'
when the 'all' option was selected for 'cell_test.max_plus_one'.

The call would look like: "virsh freecell --cellno --all" instead
of "virsh freecell --all", resulting in error:

    "stderr: error: cell number has to be a number"