Let test cases return a result
To be able to create test cases where the actual result is interesting
to save away, make it possible for test cases to return a status when
indicating a PASS status. The implementation is quite simple: The status
message will be triggered when the run() method returns a value. A quick
survey of the existing text cases in the -libvirt and -qemu suites
indicate that no test cases return a value, meaning that the fix should
not break backwards compability in the output.

The feature can be used for timing measurements (e.g. boot time, time
spent doing an action), throughput (netperf, iperf), latency (lmbench),
etc.

The field is free-text, and must fit on one row.

Signed-off-by: Jonas Eriksson <jonas.eriksson@enea.com>