virttest.utils_libvirtd: Only instantiate service if we need it
With the latest libvirt daemon control implementation,
non root runs of virt-test will fail. That happens because
the service class tries to discover which program is our
PID 1, and that needs root. So let's only instantiate the
libvirt class if it's actually used, sparing users of other
test types (qemu, libguestfs) to get a failure or message
because of that functionality.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>