virt-test: Add sphinx documentation
After some work with the virt-test wiki, it has been
converted to the sphinx format. Although many of the
rough edges and outdated docs were put into shape,
there's still a lot to be done. It's good to have
it out there and integrated early, where we can
work on it.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>