virsh.versionable_class: Fix getattr bug
If a given attribute is not on __dict__ of the original
wrapped module, just return the normal getattr behavior,
not just None. This fixes a bug that versionable_class.Manager()
was triggering in the nose unittest runner, since trying
to get versionable_class_unittest.__path__ would lead to
a None value.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>