virt: Fix transfer data over console bug in remote_commander.
There was problem with raw mode of console which wasn't so raw
as it looks like. This patch should fix this problem.
The patch using base64 coding for transferring data over console.
It could be quite easy implement new encoding in messenger when
it will needed.

data -> pickle -> encode base64 -> len,base64 ->
           -> console transfer ->
-> len,base64 -> unpickle -> data

Signed-off-by: Jiří Župka <jzupka@redhat.com>