documentation/source/conf.py: Add auto build of virt test API docs
Instead of manually updating the API rst files upon
new module additions, auto generate the .rst files
through a trick on conf.py.

With this we have reduced maintenance and always
up to date API docs.

The virt-test implementation has to be simpler than
other projects because we don't have autotest installed
in the read the doc boxes.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>