libvirt_vm: Resume vm in destroy() when vm is paused.
There are some cases need to suspend vm, and in postprocess,
we need to kill vm by vm.destroy(). But when vm is paused,
vm.destroy() will wait 240s to login it. Then find it is not
loginable and kill it with virsh.destroy().

We waste 240s at this case, we should check the state of vm
before we login it.

Signed-off-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>