staging.service: Set ignore_status for other runner such as RemoteRunner.
When I introduced the result-parser into service module, I think we will pass
a session.cmd() as a runner to service. So we should only set ignore_status
when runner is utils.run.

But later, we assume the runner passed to service should be utils.run-like, and it
should support ignore_status, such as remote.RemoteRunner. We need to set
ignore_status of all runner to True here now.

Signed-off-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>