utils_config: Remove sync() function
It's not wise to add a sync() function to restart libvirtd in
the config file class. For example, if many config file being
edited at the same time, redundant restart of libvirtd will be
introduced.

This patch remove the sync() function and the restarting activity
when calling restore(), leaving these actions to the testing
layer and concentrate on config file only in this module.

Signed-off-by: Hao Liu <hliu@redhat.com>