virttest.utils_test.libvirt: Add helper function to check command result
This helper function is an enhanced version to existing function
check_exit_status().

The original one only tests whether the exit status matches expectation.
The newer one also consider the content of the error messages if the
command is expected be fail.

Add expected fail patterns (regex) as a list for expected_fails
argument.  If test stderr matches non of these, an TestFail will emit.
If this list is empty. Command is expected to be successful. If the
command is expected be fail but the error pattern does not matter,
setting any_error to True will suffice.

If any of the skip_if list of regex matches command stderr. The test
will skip.

Signed-off-by: Hao Liu <hliu@redhat.com>