virttest.qemu_devices: Use per-image based AAVMF vars on ARM
The UEFI storage (AAVMF variables files) is bound to base image to be
able to boot the image it needs to be configured properly. This is a
very simple implementation which bounds the AAVMF file to the first
image of the VM. The same VMs would share the single AAVMF file. This
should work fine as they are only modified during installation.

The result of this patch is that when you use multiple VMs of the same
base_image, they'll start properly. In case the file is missing, it's
created and WARNING message is printed into the log file informing about
the consequences.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>