virttest.utils_test: Add option m to the tar command
Sometime the different setup in guest and host will cause the timestamp
in the future when we extract auotest package inside guest when we
try to run autotest inside guest. As this is not the test point and we
don't want this to influence our testing, add the option m to the tar
command to ignore this.

Signed-off-by: Yiqiao Pu <ypu@redhat.com>