staging/service: refactor parser for SysVInit/Systemd
The parser function looks very massive and messy, and
hasn't an unclean structure.
One function do one thing at a time and do well,
which is a golden rule.
so it's necessary to refactor them and make it tidy,
readable and easier to understand...

Modifications includes the following,
* sys_v_init_result_parser()
  |-> sysvinit_result_parser()  (keep uniform prefix)
  |-> Split it into smaller functions
      |- sysvinit_status_parser (service status)
      |- sysvinit_list_parser   (service list)
      |- raw_status_parser      (shared by Systemd&SysVInit)
      |- _ServiceResultParser.default_method (default)
* systemd_result_parser()
  |-> systemd_result_parser()  (keep uniform prefix)
  |-> The same as sys_v_init_result_parser()
    * polish the docstring and comments

This commit allows people add new feature easily,
As follow,
1) Just coding the new parser function at first
2) and then, declare this function at parser-dict[]