virttest.qemu_virtio_port: Avoid infinity hang
The socket.sendall() sends the data in an infinite loop until they are
successfully sent. When virtio_console is really hanged and the buffer
is full this hangs the test for infinity. Using socket.send('.') usually
sends the data and proceeds to cleanup. Otherwise VM is destroyed later
in this cleanup function.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>