virttest: Add bootstrap.setup()
For a long time, we've been using bootstrap.bootstrap()
before test jobs in virt test. The choice was made on
an attempt to make the process of running tests as
robust as possible - for example, users could even skip
bootstraping and things would still work.

After lengthy discussions, we reached the conclusion that
bootstrap has to be a mandatory step, and if the user skips
that, he/she will simply get an error, instead of trying to
be overly smart and do bootstraping even if the user doesn't
want it to. So, create a new setup() function that only does
the minimal necessary setup before tests, which is, restore
test images from pristine if specified in the configuration.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>